% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_to_string.R
\name{calls_to_string}
\alias{calls_to_string}
\title{Take a list of calls, return a string
Each element of the list will be converted to a string, and then
combined with the other elements}
\usage{
calls_to_string(list_of_calls, sep = " + \\n")
}
\arguments{
\item{list_of_calls}{a list of calls}

\item{sep}{a character string to separate the items of `list_of_calls` when
they are converted to a single string}
}
\value{
A length-one character vector; each element of `list_of_calls` will
be converted to a string; each string will be combined, separated by
`sep`
}
\description{
Take a list of calls, return a string
Each element of the list will be converted to a string, and then
combined with the other elements
}
\keyword{internal}
