% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{check_if_date}
\alias{check_if_date}
\title{Check if x and/or y-scales of a built ggplot2 object are date scales}
\usage{
check_if_date(built_plot)
}
\arguments{
\item{built_plot}{A built ggplot2 object, created with
`ggplot2::ggplot_build()`}
}
\description{
Check if x and/or y-scales of a built ggplot2 object are date scales
}
