% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-half-errorbar.R
\docType{data}
\name{GeomHalfErrorbar}
\alias{GeomHalfErrorbar}
\title{Half errorbars for ggplot graphs}
\description{
Geometry to draw error bars in ggplot which extend only above
and NOT below the data.  Aside from this difference, functions
identically to \code{ggplot2::geom_errorbar()}.
}
\keyword{datasets}
