% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_req_aes.R
\name{has_req_aes}
\alias{has_req_aes}
\title{Check if a call to a ggplot2 geom has the required aes}
\usage{
has_req_aes(geom)
}
\arguments{
\item{geom}{A geom call, such as returned by `geom_text()`}
}
\value{
Logical. `TRUE` if the call contains all required aesthetics.
}
\description{
Check if a call to a ggplot2 geom has the required aes
}
\examples{
library(ggplot2)
geom_a <- geom_text()
has_req_aes(geom_a) # returns `FALSE`

geom_b <- geom_text(aes(x = 3, y = 30, label = "something"))
has_req_aes(geom_b) # returns `TRUE`
}
\keyword{internal}
