% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_layer.R
\name{make_layer}
\alias{make_layer}
\title{Create a ggplot2 annotation given some aesthetics, parameters, and
facet variables + values.}
\usage{
make_layer(geom, aes = NULL, params = NULL, facets = NULL)
}
\arguments{
\item{geom}{Geom to annotate, such as "text".}

\item{aes}{Named list of aesthetics with corresponding data values, as in
`list(x = 3, y = 30, label = "A label")`.}

\item{params}{Optional. Named list of parameters for geom,
such as `list(colour = "black")`}

\item{facets}{Optional. Named list of facets; the name is the variable and
the value if the level, such as `list(cyl = 4)`.}
}
\value{
A call
}
\description{
This function is based on ggplot2::annotate(). The two main differences are:
it returns an unevaluated function call rather than a ggproto object; and
it can handle limiting annotations to (specified) facet level(s).
}
\examples{
library(ggplot2)

base_plot <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  facet_wrap(~cyl)

my_annot_call <- make_layer("text",
  aes = list(x = 3, y = 30, label = "A label"),
  facets = list(cyl = 6),
  params = list(col = "red")
)

my_annotation <- eval(my_annot_call)

base_plot + my_annotation
}
