% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_unit.R
\name{safe_unit}
\alias{safe_unit}
\title{Functions to safely work with grid::unit() and grid::arrow() in Shiny
If an input is NULL, NULL is returned; otherwise return a unit/arrow object}
\usage{
safe_unit(x, units)
}
\description{
Functions to safely work with grid::unit() and grid::arrow() in Shiny
If an input is NULL, NULL is returned; otherwise return a unit/arrow object
}
