% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_plot.R
\name{selected_plot}
\alias{selected_plot}
\title{Parse text selected in the RStudio IDE as a ggplot2 plot}
\usage{
selected_plot(text = rstudio_selection())
}
\arguments{
\item{text}{String to parse as ggplot2 object;
default is `rstudio_selection()`}
}
\value{
A ggplot2 plot
}
\description{
Parse text selected in the RStudio IDE as a ggplot2 plot
}
\keyword{internal}
