% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_layers.R
\name{combine_layers}
\alias{combine_layers}
\title{Take lists corresponding to several annotations, combine annotations
that share parameters and geom, and return a list-of-lists.}
\usage{
combine_layers(lists)
}
\arguments{
\item{lists}{List of lists to combine. Each sub-list should have four elements:
\itemize{
  \item{`geom`}{length-one character vector such as `"text"`, or `"point"`}
  \item{`aes`}{named list containing variable-value mappings,
  such as `list(x = 3, y = 40)`)}
  \item{`param`}{named list containing parameter names and values, such as
`list(colour = "black")`}
  \item{`facets`}{named list containing facet variable-value pairs, such as
  `list(cyl = 4)`}

}}
}
\value{
A list of lists. Any supplied lists that share a geom and parameters
will be combined, so the length of the returned list will not necessarily
equal the number of lists supplied to `lists`.
}
\description{
Take lists corresponding to several annotations, combine annotations
that share parameters and geom, and return a list-of-lists.
}
\examples{
library(ggplot2)

layer_1 <- list(geom = "text",
                aes = list(x = 3, y = 30, label = "Some text"),
                facets = list(cyl = 4),
                params = list(colour = "red"))

layer_2 <- list(geom = "text",
                aes = list(x = 4, y = 35, label = "Some other text"),
                params = list(colour = "red"))

layer_3 <- list(geom = "point",
                aes = list(x = 3, y = 40))

layer_4 <- list(geom = "text",
                aes = list(x = 4, y = 45, label = "Some more text"),
                params = list(colour = "black"))

lists <- list(layer_1, layer_2, layer_3, layer_4)

# combine_layers() combined layers 1 and 2 as they share a geom and params.

layers <- combine_layers(lists)

# The resulting list can be used to create ggplot2 annotations

annots <- purrr::map(layers,
                       ~make_layer(geom = .x$geom,
                                   aes = .x$aes,
                                   params = .x$params) \%>\%
                         eval())

ggplot2::ggplot() +
  annots

}
