% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggannotate.R
\name{ggannotate}
\alias{ggannotate}
\title{Interactively annotate a ggplot2 plot}
\usage{
ggannotate(plot = last_plot())
}
\arguments{
\item{plot}{A ggplot2 object. Default is `ggplot2::last_plot()`.}
}
\description{
Launches an interactive Shiny gadget that lets you point and click
on a ggplot2 plot to add annotations, then generates the code to
reproduce them.
}
\examples{

\dontrun{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()

ggannotate(p)
}

}
