% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggannotate.R
\name{ggannotate}
\alias{ggannotate}
\title{ggannotate}
\usage{
ggannotate(plot = selected_plot())
}
\arguments{
\item{plot}{A ggplot2 object.
If blank, your current selection in RStudio will be used.}
}
\description{
ggannotate
}
\examples{

\dontrun{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()

ggannotate(p)
}

}
