% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggannotate.R
\name{ggannotate}
\alias{ggannotate}
\title{ggannotate}
\usage{
ggannotate(plot_code)
}
\arguments{
\item{plot_code}{A string containing code to construct a ggplot2
object. If blank, your current selection in RStudio will be used.}
}
\description{
ggannotate
}
\examples{

\donttest{
ggannotate("ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point(col = 'orange')")
}

}
