% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggannotate.R
\name{ggannotate}
\alias{ggannotate}
\title{ggannotate}
\usage{
ggannotate(plot = last_plot())
}
\arguments{
\item{plot}{A ggplot2 object. Default is `ggplot2::last_plot()`.}
}
\description{
ggannotate
}
\examples{

\dontrun{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()

ggannotate(p)
}

}
