% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggannotate.R
\name{ggannotate}
\alias{ggannotate}
\title{ggannotate}
\usage{
ggannotate(plot)
}
\arguments{
\item{plot}{Either a ggplot2 object, or code to construct a ggplot2
object. If blank, your current selection in RStudio will be used.}
}
\description{
ggannotate
}
\examples{

\dontrun{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
       geom_point()

ggannotate(p)
}

}
