% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_layer.R
\name{make_layer}
\alias{make_layer}
\title{Generate a call to ggplot2::layer() given some aesthetics and parameters
Based on ggplot2::annotate()
Can handle limiting annotations to (specified) facet level(s)}
\usage{
make_layer(
  geom,
  x = NULL,
  y = NULL,
  xend = NULL,
  yend = NULL,
  label = NULL,
  params = NULL,
  annotate_all_facets = FALSE,
  facet_var1 = NULL,
  facet_level1 = NULL,
  facet_var2 = NULL,
  facet_level2 = NULL,
  ...,
  na.rm = FALSE
)
}
\arguments{
\item{geom}{Geom to annotate, such as "text".}

\item{x}{Variable to map to x aesthetic}

\item{y}{Variable to map to y aesthetic}
}
\description{
Generate a call to ggplot2::layer() given some aesthetics and parameters
Based on ggplot2::annotate()
Can handle limiting annotations to (specified) facet level(s)
}
