% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_layer.R
\name{make_layer}
\alias{make_layer}
\title{Generate a call to ggplot2::layer() given some aesthetics and parameters
Based on ggplot2::annotate()
Can handle limiting annotations to (specified) facet level(s)}
\usage{
make_layer(
  geom,
  aes = NULL,
  params = NULL,
  facet_vars = NULL,
  facet_levels = NULL
)
}
\arguments{
\item{geom}{Geom to annotate, such as "text".}

\item{aes}{List of aesthetics with corresponding data values, as in
`list(x = 3, y = 30, label = "A label")`.}

\item{params}{List of parameters for geom, such as `list(colour = "black")`}

\item{facet_vars}{List. The names of variables used to facet the plot,
such as list("cyl").}

\item{facet_levels}{List. The levels of variables you wish to annotation, such
as list(4).}
}
\value{
A call
}
\description{
Generate a call to ggplot2::layer() given some aesthetics and parameters
Based on ggplot2::annotate()
Can handle limiting annotations to (specified) facet level(s)
}
\examples{
library(ggplot2)

base_plot <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()

my_annot_call <- make_layer("text",
  aes = list(x = 3, y = 30, label = "A label"),
  params = list(col = "orange")
)

my_annotation <- eval(my_annot_call)

base_plot + my_annotation
}
